feat: implement selection caching and effective selection retrieval for improved Omnibar command handling
